tests/requirements.txt
transformers>=4.30.0
pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
pytesseract>=0.3.10
